requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
exceptiongroup==1.1.1
fastapi==0.95.0
future==0.18.3
h11==0.14.0
idna==3.4
imagesize==1.4.1
outcome==1.2.0
pycparser==2.21
pydantic==1.10.7
pylivestream==2.0.0
pymediainfo==6.0.1
PySocks==1.7.1
#python-magic==0.4.27
#python-magic-bin==0.4.14
selenium==4.8.3
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.26.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
wsproto==1.2.0
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.23.3

